支出CRUD实现

+++ app/src/main/java/com/example/account_book/network/model/NetworkTransaction.kt
@@ -3,7 +3,7 @@ package com.example.account_book.network.model
 import com.google.gson.annotations.SerializedName
 
 data class NetworkTransaction(
-    @SerializedName("id") val id: Int? = null,
+    @SerializedName("id") val id: Long? = null,
     @SerializedName("amount") val amount: Double,
     @SerializedName("categoryId") val categoryId: Long,
     @SerializedName("remark") val remark: String?,
